requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
# transformers
# SentencePiece
langchain-pinecone
langchain-community
langchain-openai
langchain-experimental
-e .